requirements-dev.txt
# main
flask
pyTelegramBotAPI
flask_redis
haversine
pillow
# dev
pre-commit
flake8
flake8-docstrings
black
pylint-flask
